Use own ModelView matrix
The camera module in the opengl directory provides an alternate
camera implementation for pyFormex. It is intended to replace the
old camera module in future. It will only use OpenGL 2+ and OpenGLES
compatible features.
The current version has the ModelView matrix completely implemented
in our own code.

To use it, you have to check out the opengl branch, and still use the
--opengl option on startup. In future we may make the --opengl option
default in the opengl branch.