Fix the gts commands output on invalid options.
The long command options for the gts commands do not work on my system.
Instead they produce invalid short options, causing the printing of a message
with the advice to run 'gtscommand --help', which of courses just produces
the same message if you run it. Therefore I changed the message to
'gtscommand -h', because the short options should always work.

I have no idea why the long options do not work.