trisurface.check has been improved
runCommand has been replaced with utils.system becuase
this function does not operate on the surface, but returns info about
 the surface. When using runCommand it was just raising an error,
without returnig the info.
Also,verbose option has been added: in gtscheck the -v option is needed
to return statistics of the surface, both when sta=0 and sta!=0 (so it is
not an error message).