Changed Connectivity.connectedTo and Connectivity,testDuplicate
The extra argument return_mult has been changed to return_multiplicity
and its working is different: it makes the multiplicities to be
returned *in addition to* the normal return values. This is to be
consistent with other uses of return_... arguments.
See e.g. Connectivity.testDuplicate.
Question: should we also return the multiplicities in a dictionary,
like in testDuplicate?

The dictionary returned by testDuplicate now has integers as keys
instead of strings.