Fixed a bug in vmtk_itf.remesh
A new argument 'elementsizemode' is added to specify which metric is used for remeshing (edgelength or area).
Previously, the target edgelength was passed to the vmtk command, but elementsizemode was not specified.
The default settings (-elementsizemode area -area 1) were therefore applied by vmtk, leading to unexpected results.
I also added an extra argument 'aspectratio` to specify the quality of the triangles.