Reactivate the interactive Python console
The interactive Python console is now in a separate module pyconsole.py.
It can be activated by setting the  interpreter variable in Settings,Startup.
Beware that the console is an experimental feature and many things need
to be improved and checked. There may be subtle things that break other
pyFormex features. You have been warned!
Nevertheless, if you find it useful, go ahead and use it, and help us to
test and improve it.