Opengl2: Specular color;  Drawable class
The vertex shader now also has specular color, defined by 3 attributes:
- specular: global intensity of specular color
- speccolor: rgb value of the specular color
- shininess: surface shininess

The rendering of the objects has been moved to a Drawable class, and
the Actor class contains a list of drawable. These allows to use
multiple rendering actions for one Actor. This can e.g. be used to
render front and back plane, or surfaces and edges (*wire modes).