Add generic element names to Isopar
The isoparametric transform class Isopar now supports generic element
type names. For example an eltype 'lag-2-2' will use a lagrangian grid
of 3 points in x and 3 points in y direction. This corresponds with the
'quad9' element. Note that the eltype coding uses the degree of the
polynomials, which is one less than the number of points. Another example:
'bor-6-6' will use the circumference of a square with 7 points along each
side.
The 'exponents' function has been changed in the same way (using degree).

The traditional element names like 'quad9' can still be used, but are now
aliases for the generic names. We recommend to used the generic names,
certainly for high orders, for unequal orders in different directions,
and for names that are not defined in elements.py.

Also, the static data table has now been replaced by an ondemand filled
table.