Add multivariate Polynomial class
The Polynomial class represents a poynomial in n variables.
It is currently represented by an array with the exponents, but
a string expression in human readable form can be obtained.
The polynomial is now used in the Isopar transform. Some functionality
of isopar.py has been moved to polynomial.py.