Allow selection of the OpenGL version
A command line option has been been added to set the prefered OpenGL
version. For example --opengl=3.1 will select that version.
The Help->OpenGL Format option now prints the current version and all the
supported versions.

Note that the master branch is only using 1.0 features and this is probably
what most installations will be using by default.
The new opengl engine will target version 3.3 or higher, possibly with a
fallback to some lower version for old hardware.

This new option is primarily intended for the developers of the new opengl
engine. That's why it is not a configurable option (yet).
Normal users could possibly use it to get some performance improvement
or to work around some OpenGL hardware/libraries problems.