Fix bug #39098: TriSurface.areas() and Mesh.volume()
The TriSurface.areas attribute has been renamed to _areas. The facetArea
method, returning the area of all facets, has been renamed to areas,
for consistency with Mesh.areas.
facetArea can still be used, but is now deprecated.

Mesh.volume() now computes the volume by converting the border to a
surface.