Added two functions to calculate distances to surfaces to vmtk_itf.py
vmtkDistanceOfSurface calculates the distances from one surface to another.
vmtkDistanceOfPoints calculates the distances from a set of points to a surface.

The function's names start with 'vmtk' and the functions are currently not
installed as TriSurface methods because a similar method already exists for the
TriSurface class (distanceOfPoints). However, the vmtk functions seem to be faster
and, unlike the current TriSurface method, do not result in memory errors for large
geometries.