Add a dynamical callback function to some InputItems
The InputBool and InputColor now have an option 'func', much like was
already the case with InputFSlider. It specifies an optional function
that will be called whenever the current value of the input item changes.
For the InputColor, this function is also inherited by the color dialog
that is popped up when clicking on the color input field.
The function gets the input field as a parameter. The current value can be
obtained from it. Optional data can also be passed as the 'data' attribute
of the field.

This 'func' option should probably be generalized for all (or most) of
our InputItems. It allows some powerful effects, like dynamically changing
the colors of displayed objects. Especially the new OpenGL engine will
benefit (as illustrated in example3 in the opengl branch).