Create single project Description file
pyFormex was until now differntly described on several places: the Description
file in the top directory, the 'long_description' in the setup.py file,
the 'About' section on the website home page. They all used somewhat
different wording. Now we have unified into a single source (the Desciption
file) that is then included in the setup.py and in website/index.rst.
The Description file is in 'rst' format, but we try to keep the number
of ReSt features to a minimum, so it can be easily read as simple text too.

The HOWTO-dev.rst file was renamed to HOWTO.dev to allow easy git
attribute setting on (other) .rst files.