opengl2: implement more color modes
Using per element color, using indexed colors with a colormap and the
pyFormex default color mode (property number as an index) now also work
in the new opengl engine.
You still need to add color='prop', because the defaults for the new draw()
function are not yet applied. An example like Palette displays fine, because
we added color='prop'.