Revert the change made by previous commit
Revert the playScript function.
Major changes to the basic functionality should not be done so lightlty.
They should be discussed, widely tested, and only then implemented in the
mainstream.
The console has its own environment for executing scripts, and that
should be used. The playScript functionality was never intended to be
adresses from multiple places.