Move the Mesh/Formex setProp to Geometry
Moved the setProp method up to the Geometry base class will m ake it
easier to implement algorithmes to set the property numbers in a
consistent way.

Removed the layerprop parameter from Mesh.connect. It is contrary to the
current pyFormex logic: when generating new geometry out of old, the new
elements inherit the property from their ancestor.
If you want other properties, they should be set afterwards, but in
general fashion to be available to all objects. The move of the setProp
method facilitates this. One example is already implemented:
setProp('range').