Backport wiremode toggle from opengl branch
The new method for setting/toggling the wire mode overlay drawing
have been backported from the opengl branch to the master branch.
This has no practical consequences for the user, because the
extra wire modes are not implemented in the master branch.
But it allows us to keep the two brancehes more in sync.