Fixed bug in gui.actors.GeomActor.draw while handling wireframe mode
the class does not have default value self.wire. so while checking the value
self.wire raised an error in case the self.wire did not exist yet.
now a check if wire attribute is performed.if true and the wiremode differs
from the canvas one the attribute is removed to be handled as the attribute wire did not exist