Opengl2: add Camera.toNDC
Camera.toNDC converts a set of world coordinates to normalized
device coordinates. These coordinates range from -1 to +1 for
points falling in the viewable frustum (and thus also inside
the canvas viewport).
Furthermore, a pickwindow may be specified as 4 integer values,
specifying a subregion of the canvas viewport. In that case, points
rendered inside that window will be mapped in the -1 to +1 range.

With this function any kind of mouse selection by a rectangular
area will become a piece of cake.