removed clean=True parameter in vtk_itf.writeVTP
clean=True performs in vtk something similar to a combination of renumber, removeDuplicates,
and fuse and compact function of pyformex on the mesh but does not change the
props order of the mesh. so it should not be called by in this function.
the user should previously perform a cleaning outside i.e using convert2VPD(...clean=True)
and then convert2VPD to check i the coords/ elems of the mesh are changed.
currently there is not a way in vtk_itf to follow the changes performed by the
vtkClean function on the elems,coords