Allow add/remove of Collection to/from Collection
You can now add/remove data in the form of a Collection to/from
another Collection.
Collection is used as an accumulation buffer to store entities
from the current pick buffer. In the master branch, this contains
pure numerical data. But in the new opengl branch, the current pick
buffer is already a Collection. This changes allows to work with both.

This commit also makes a minor change to viewport.py to allow the
use of a Collection as pick buffer.