opengl2: modified the controllers of the webgl gui.
The control panel has a folder for each pyFormex actor and its
controllers act on the actor's drawables. The attributes (visibility,
opacity, control) can thus be modified separately for each drawable
(e.g. front-facing triangles, back-facing triangles, edges, wires).

An option should be added to include controllers acting on the actor
(e.g. visibility), which are passed to the drawables.