Change the use of color/alpha attributes
Setting color and alpha values as an attribute of the object was
introduced earlier to simplify the generation of WebGL models.
The color/alpha attributes were honored by the draw() function.
Now these attributes have been moved to the .attrib attibute of
the object. This means that instead of 'F.color=red' you should
use 'F.attrib.color' = red.
The honoring of these colors is no longer done by the draw()
function, but by the GeomActor constructor.
The webgl exporter and the WebGL example have been adapted
accordingly.

The use of the intermediate .attrib might be left out again in
future, after we switched to the new opengl branch.