Add _opengl2 attribute to Examples
In order to help with the transition to the new opengl2 engine,
we added two new attributes to the example scripts:
_opengl2: if True, specifies that the example runs (but may still have
defects); if False, means that the example does not run (causes an
exception); if unspecified, means that the example was not tested yet.
_opengl2_comments: a (multiline) string with any comments about aspects
that still need to be fixed, or observed differences with the old engine.