Fix props in Mesh.splitByConnection (bug #40149)
splitProp has got an optional prop argument, to split on a given set of
property values, without having to insert the values into the object.
This is used in Mesh.splitByConnection to solve bug #40149.
Note that splitProp was moved up to the Geometry class, thus working for
Mesh and Formes (and others) at the same time.
Also added Geometry.toProp(prop) to convert an argument to proper prop
set (setting length + int type).