Revert changes from 4f80577 to writeTetMesh
Rationale: All the functions in the tetgen plugin take a full filename
as parameter, including an extension. The docstrings also say so.
Changing this should involve the same change in the whole plugin and
that would require a lot of incompatibility warnings.
The 4f80577 behavior can easily be got by doing a call
writeTetMesh(fn+'.node',...)