Fix memory leak (bug #39950)
Added a __del__ method to gui.drawable.Drawable. The method cleans
up the OpenGL display lists when the Drawable is destroyed.

Note that this relies on __del__ being called, which is not guaranteed
by Python, but in normal applications it should suffice to remove
the memory leak.

Also note that this fixes the bug for the old OpenGL engine in the
master branch. The new engine (in the opengl branch) should still be
fixed.