Fix 2.7+ syntax and cleanup
The vtk_itf contained some syntax that is only available since Python 2.7.
Because pyFormex should still be compatible with 2.6, these needed to be
changed. The construct involved implicit loops inside a dict definition.
In 2.6 that only works with list, not dicts.
Thus the 2.7 syntax
d = { k:v for k,v in zip(klist,vlist) }
could be changed to work in 2.6 to
d = dict( [ (k,v) for k,v in zip(klist,vlist) ] )
But there is an easier way: the dict constructor allows
direct initialization with a zipped list:
d = dict(zip(klist,vlist))

Also, the code has been cleaned to adhere better to the pyFormex coding
style (assignment statements have a space around the = sign), and the
docstring formatting was fixed (should be valid ReST!)