Preparing release 0.9.1
Some finetuning for the next release.
When upgrading from 0.9.0 to 0.9.1, the user configuration is moved
automatically from ~/.pyformex/pyformexrc to ~/.config/pyformex/pyformex.conf.
Now we also link back to the old locations, so that the older version will
still run if the user has such a need.

Remove the deprecated formex.mpattern function.