Changed the default palette
Color number 0 is now darkgrey, and color 8 is black.
This gives more consistency as colors 0-7 are now lighter versions of
colors 8-15. See example Palette. The definition of darkgrey has been
changed to make it a little darker (0.4 instead of 0.5).
The change also allows the required light intensity more homogeneous
over the first 8 colors.

The default drawing color is still black. We may consider to also change
it to darkgrey.

Notice that the default palette is set in the config file; the palette
in colors.py is still not used.