Prepare the master switch
In preparation of switching the master branch to the opengl development
branch (the old master will remain available as the 0.9 branch),
we have added two command line switches: --gl1 and --gl2, to select
either the old or the new engine.
To ease the switch for users following the master branch, we have made
the OLD engine the default. The new engine will only become default
when it becomes nearly fully functional.
Users of desktop buttons to start pyFormex can create new starters
with the install-pyformex-dev-desktop-link, e.g. one with GL1 ans one
with GL2 as arguments. You will have to change the properties of the
starters to include the --gl1, resp. --gl2 in the command line.