Bumping pyFormex major version to 1.0.0
Bumping the version is required after a release. But as now will now
make the new opengl engine development branch the master branch,
and leave the old 0.9 branch in maintenance status only, it is
appropriate to not only bump the minor version, but bump the major
version instead. Thus, the next release will be version 1.0.0.