Switch the master branch (IMPORTANT)
The master branch has been switch to the opengl development branch.
The old branch remains available as the branch 0.9, for maintenance.

For most user following the master branch, there should be no important
consequences. In the new master branch, you can still select the old
opengl engine, and it is even the default.
Selection is by using --gl1 or --gl2 command line option.
Users can have a peek at the new engine by using the --gl2 option.
With time, support for the old engine will be reduced and finally
dropped, and the new engine should become the default.

The master switch makes it easier to maintain the repository, as users and
opengl developers have now again the same source tree.