Restructure script mode; Fix bug #39379
All execution of scripts is now done over a second Python interpreter,
(i.e. not the one running the pyFormex main program/GUI.
The script interpreter is constructed at startup, and passed to the
interactive console, it that is used. Thus scripts get executed in
the same way regardless of activating the console.

The backdraw is that we can only exit from the script by raising a
SystemExit. Therefore exit() now will raises a SystemExit, which  is
caught by the script executor (we do not want to exit pyFormex).
Thus exit() now works also with the console activated (fixes bug #39379).
But the exit(all=True) to exit completely from pyFormex is no longer
possible. I doubt however that anybody uses this. If someone needs it,
we can solve it by adding some exitcode to the script's globals.