Allow easy addition of app paths
As it has increasingly become clear that executing in script mode
does not work with multiprocessing (and presents other problems too),
users are invited to go with the 'app' model for anything but simple
linear scripts. A good rule is to use an app as soon as you make
use of class or function definitions. Certainly never use any
multiprocessing functions in a script. By preference use only the
pyFormex multi.multitask for your parallel processing: it will issue
a warning if accidentely you switch to 'script' mode.

The drawback of loading your applications as an 'app', is that they
should be placed in a path listed in Python's sys.path. This can be
configured from the App->Configure paths menu. But when you use the
script<->apt button to switch from script to app, you often get an
error message about the path of the script not being in the configured
directories. Now, an option has been added there to automatically add
the required path with a simple click. Hey, are we user friendly, or not?