Add light config to render settings
The Settings->RenderParams dialog has been extended with the light
settings for 4 lights. The lights can be changed individually, not
only in their intensity, but also in the color of the light they
emit in each of the types: ambient, diffuse, specular.
Note that global ambient is (currently) always white. To set color
ambient, dim the global ambient and let the lights emit ambient
color.

Todo: the rendering should be updated real-time when changing the
light values.
Note: some of these (and any future) functionality may only be available
in the gl2 engine.