improve remesh and vmtkDistanceOfSurface in vmtk_itf
- To remesh a TriSurface the user can specify the regions to be remeshed
using the new option `includeprop`. The rest of the surface will not be
remeshed. The option `inlcudeprop` is not compatible with `excludeprop`.

-In vmtkDistanceOfSurface all temp files to write and read are now
vtp files instead of stl and dat files. The advantage is that vtp files
are smaller than stl and that the resulting distance is also written inside
a vtp file, resulting in much faster writing/reading operations. This
is also a first step towards the use of vmtkDistanceOfSurface as a python
function instead of an executed external program.