whereProp and (c)selectProp moved in Geometry
-whereProp has been moved from Formex to Geometry,
to make it working also for Mesh. Note that
whereProp implementation has been rewritten
in a more compact way, still holding previous
functionality.

-withProp (in Formex and Mesh) and withoutProp
(it was only in Mesh) have been renamed as selectProp
and cselectProp and both moved into Geometry.

NB: as in the previous implementation, if a Mesh/Formex
has no prop, (c)selectProp returns the full Mesh/Formex,
while whereProp returns an empty array. Would not it be
safer to raise an error in whereProp and (c)selectProp?