Activate drawing attributes in gl2
Most of the draw() parameters of the old engine are now also active in
the new gl2 engine. Thus, default draw will now also be with color='prop'.
The 'clear' parameter is stored as 'clear_' though, because a Dict storage
dies not allow a key 'clear'. The access function (setDrawOptions) still
allows using 'clear'.