Relax script locking.
Script locking for the 'app' model has been relaxed. You can now tell
runApp to not check the lock and to not create a lock. This allows one
app to execute another app. E.g. as in the following code:

from gui.draw import *
def run():
    for a in [ 'BarrelVault', 'Baumkuchen' ]:
        runApp('examples.%s' % a,lock=False,check=False)
        sleep(1)

This will execute the two specified examples.

Currently default stays to create and use the lock, but this might change
in future.