Revised utils.system (bug #40566)
Avoid an exception when trying to execute a non-existing command.
See also https://savannah.nongnu.org/bugs/index.php?40566#comment13
Instead of an exception, the Process attribute 'failed' is set True,
allowing the detection. Also, the returncode is set to 127, which is
also the code returned by the bash shell on a non-executable command.

The verbose argument of the runCommand function has been moved up to
utils.system, and a new argument raise_error has been added to flag
whether errors in the command should cause an exception.

Finally, added a new command() function intended to replace runCommand
for user scripts.

To summarize, we now have:

utils.system: the low level, to be used by preference in pyFormex code.
It is silent, and raises no exceptions (by default, can be changed by
the parameters).

utils.command: is verbose and raises exceptions on errors other than a
timeout. This is better for user application scripts (which normally
should not continue if something went wrong).

utils.runCommand: this is like command, but with shell=True and returns
a sta,out tuple. This is kept for compatibility, but should be replaced
with command or system and the proper arguments. Will probably be
deprecated in future.