gtsset: check for specific errors in P.err
Screen P.err resulting from gtsset in order to find specific errors.
Added an error case:
intersection of surf1 and surf2 is not a closed curve.
This is always the case when a boolean is performed on
not intersecting surfaces. In this case instead of an raising an error,
 a string is returned in order to alert the user. As pointed out
in sr #108442 there can always be an other reason for this error.
The user should recongnize this specific event.