Reverse commit 0b4b6d9e
The error parsing introduced by 0b4b6d9e has been removed, as it is
considered needless cruft. It does not make sense to scan all error
messages and replace them with another error message. The best thing
we can do on an error is either
- let the thing continue until it hits an exception. The user can capture
  it and act accordingly
- capture the error ourself and give information to the user by printing
  some message and/or returning a special error code.
We just have to make sure we do not return valid but incorrect results.

Therefore I throw out the error detection code from gtsset.
I did not reverse to the original (which probably raised an exception),
but let the function print the error and return with a None value instead
of the TriSurface or curve (maybe we should change the returned curve
to a PolyLine instead of a Formex?).