Extend utils.Process functionality
The parameters stdin, stdout and stderr of utils.Process can now also
be filenames instead of open files. This simplifies the use of files
as input/output of commands, avoiding the use of redirection symbols
and thus the need for a shell to launch the command.
Thus, you can now use
  utils.command("cmd...",stdin="file.in",stdout="file.out")
instead of
  utils.runCommand("cmd...  < %s > %s" % ("file.in","file.out")

As a test and example, the TriSurface methods coarsen and refine have
been changed in this manner.