Sanitize warnings; fix some bugs.
All warnings with utils.warn have been moved to messages.py, and use a short
mnemonic string in the warn call. This helps a lot with the warnings
filtering, which sometimes does not work correctly with long, complex
strings.

* Henceforth all utils.warn calls should use this technique: a short
mnemonic string (only alphanumeric and '_') in the call, and the full
text in messages.py.

In order to allow variable text warnings, the utils.warn function now
has an optional 'data' argument. If data are specified, the printed
message will be message % data, so you have to make sure you put the
correct number of formatting fields in you message.

This commit also fixes some bugs. The most important is the use of the
exit() function in an app: it will now correctly terminate the running
of the app, but not terminate pyFormex.

All the exceptions defined in script.py have been removed, since the
SystemExit is now the only exception that can terminate the execution
of a script.