Removed Mesh.connectionSteps
As Mesh.connectionSteps does indeed reproduce the same functionality
as a frontWalk, its functionality has been removed and a warning is
displayed pointing to the FrontWalk example.

The new FrontWalk example provides somes example walks over a Mesh.
Walks 2 and 3 show how to start a walk from both elements and nodes
at the same time. Since in most cases a walk will start *either* from
elements or nodes, and not from both, the simple cases of walks 0 and 1
will be what is generally used.

Note that the return value of a frontWalk is different than that of
connectionSteps: frontWalk returns a single array with a number for
each element, specifying the step in which it was passsed.