Fix a bug in ArrayModel
Editing of numerical arrays in a table using the InputDialog was not
possible, because of a bug in the ArrayModel. Validating the field
data did not work. I think that it was made for the old PyQt4 API.
The bug is fixed now for use with PySide. I have not tested the
PyQt4 with the new API.