Fixed extrude methods
After the recent change of the Mesh.extrude arguments, the examples were
not adapted and neither was the Formex.extrude method.

Now Formex.extrude uses the same order and meaning of arguments as
Mesh.extrude, and all examples have been adapted to the new args.
Furthermore, an extra check was added to Mesh.extrude to detect some
obvious errors against the new argument order. In such case, a warning
is displayed, even if the user has flagged off the default message
displayed on every extrude call.