Add varray module
The varray module defines the Varray class. This class provides 2D
integer tables with varying row length, such as the inverse of a
Connectivity table. For large models, the Varray class uses a lot
less memory and is faster too. See the docstring of the module for
details.

We intend to replace the inverse tables with a Varray wherever
possible. However, because the Varray is not an ndarray, it misses
a lot of methods that exist for the current inverse tables.
We will proceed step by step, and will probably need to add some
more methods to the Varray class.

New code needing inverse tables can better be directly written to
use the Varray. The varray module contains an inverseIndex function
that can replace the arraytools.inversIndex, but produces a Varray.