Improved inverseIndex
arraytools.inverseIndex now uses varray.inverseIndex to faster
compute the inverse index. The old inverseIndex functions has
been retained for the (unlikely) case that something would go
wrong with the new function.

varray.inverseIndex now also accepts a Varray as input. It is
converted to an ndarray first.

varray.Varray accepts an ndarray with negative values as data.
The negative values will silently be ignored.

varray.Varray has a convenient method inverse(), which uses
the inverseIndex function. The inverse of the inverse yields
the original, possibly with the exception of the order of the
numbers in the rows.