Major changes in vtk_itf for convertFromVPD and readVTP
readVTP is deprecated and has been renamed readVTKObject. In future it migth contain
other vtk data reader (not sure if a general reader already exist).
All the fields readers are moved from readVTP to convertFromVPD (adding a check
for the array data type) which now returns also field cell and poins data array.
readVTKObject works as readVTP the only difference is that now
the coords output is included in the list [coords,polys,lines,verts] to keep
consistency with output of convertFromVPD
As not all the data are present in all vtkdatatypes a check on the data type is
performed using GetDataObjectType() before trying to store the data array. A commented list
of all datatype and their integer reference has been addaed in the plugin.