Rename/add front methods
In both Connectivity and Adjacency, the 'front' method (which was
called 'frontFactory' before, has been renamed again to 'frontGenerator'.
I think that is the proper name, telling what it does. It also leaves the
name 'front' available for other purposes, and that is just what we have
done: both classes now have a 'front' method that will return the numbers
of the elements in the very first front reached from the startup.
It makes sense to add such a method because if you only want a single step
and never intend to walk any further, these first front elements can be
obtained a lot cheaper than by initializing a frontWalk and only perform
one step.

The front method has an option 'add': if True, the returned element set
also contains the starting elements.