More Varray goodies
A Varray can now be initialized with another Varray. Thus va = Varray(a)
will work whether a is an ndarray of a Varray.
Also, a Varray can be changed inplace by the data of another Varray.
  va1, va2 =  Varray(a1),Varray(a2)
  va1.replace_data(va2)
Now va1 will be equal to va2. This method is intended for internal use.
The normal user should use:
  va1 = Varray(va2)