Work around a bug in gtsinside
Due to a bug in the testBbox algorithm used in pyformex_gts.inside,
erroneous results were reported in some cases. Therefore, the clipping
of the surface in directions orthogonal to the shooting direction has
currently be deactivated. This has a time penalty, but at least the
results are correct.
We will reintroduce the clipping when the origin of the bug has been
found (and solved).