Run gtsinside in parallel
puformex_gts.inside now has an option multi=True, which will run the
gtsinside command for the three shooting directions in parallel,
allowing an almost 3x speedup. This option has also been added to
TriSurface.inside, but will only have an effect with method='gts'.

Note that the Inside example needs to use multi=False, because it
starts multitasking itself (not running the commands in parallel,
but splitting the data into blocks). And the functions executed
by multitask should not run multitask themselves!