Restructure Python3 development
Some files intended for a Python3 version of pyFormex have been
streamlined and put in the normal pyformex tree.
Detection of the Python3 version is now done in pyformex/__init__.py,
to make detection work on users that have installed Python3 as the
default (a growing number).
The script to (force)start the Python3 version is now called pyformex3
and remains in p2to3 directory. It will now also work when executed
from elsewhere.