Allow set literals; require Python 2.7
pyFormex now requires at least Python 2.7. Therefore we can now allow
the use of set literals in pyFormex code, which is much cleaner than
using the corresponding set() constructor. All set() constructs were
changed to set literals: set([1,2,3]) becomes {1,2,3}.

The requirement for Python 2.7 should not pose a problem: the previously
required 2.6 dates from 2008, 2.7 was released in 2010. There are
probably not a lot of installations left that default to 2.6.