Apply the zip fixer
All the code has been changed to accept the future style of the zip
function: zip() returns an iterator rather than a list. In case where
you use zip in an iteration: for ... in zip(...), nothing needs to be
changed. Also creation of a dictionary from an iterator rather than a
list works: d=dict(zip(keys,values)). But when you explicitely want
to store the list, you need to use list(zip(...)).

Since it is our aim to support in future both Python2 and Python3 from
the same source code, we have added a line
  from future_builtins import zip
to all pyFormex source files. That way you will be forced to use the
new zip style.