Fix all map functions
Like zip, in Python3 map returns an iterator, not a list. In pyFormex, map
was almost exclusively used to get the list as a result. Thus, instead of
map(func,alist), we need to explicitely write list(map(func,alist)) to get
the list.
An alternative is to use a list comprehension: [ func(i) for i in alist ].

While I like the map function, the construct with a list around it is
rather distracting, so I prefered to convert all maps to list comprehensions.
I did not add the from 'future_builtins import map' to force the future
version of map, so for now you can still use the normal map, though
I recommend the usage of list comprehensions.