Major cleanup started
As one of the last phases of cleaning the pyFormex code so that it
is accepted by Python3, we need to transform all (implicit) relative
imports to either explicit relative imports or absolute imports.
Thus, where we have now in a module pyformex.plugins.trisurface:
  import geomtools
  import inertia
this should become:
  from pyformex import geomtools
  from pyformex.plugins import inertia
when using absolute imports, or when using explicit relative imports:
  from .. import geomtools
  from . import inertia
I think the best practice is using absolute imports.

This commit starts implementing this. We will have to go over evey module
as there is no automatic fixer for this (the automatic fixer converts to
explicit relative links, and can not do everything correctly, because of
our sloppy style in the past).

You can have a look at gui/guimain.py for an idea of how to do this.
Some style guides insist on putting one import per line, but I have
the feeling that using a list is also acceptable:
  from pyformex.gui import (
      signals, QtCore, QtGui,
      menu, cameraMenu, fileMenu, appMenu, prefMenu,
      toolbar, canvas, viewport, guifunc, draw, widgets, drawlock, views,
      )

We will convert a few modules at a time, so that we can detect problems
in a limited scope.