Allow unicode filenames in geomfile
For some as yet unclear reason, the filename dialog seems to return
unicode strings instead of str. This is not a bad thing, as importing
files that have unicode names might be an interesting feature.
However, that made the GeometryFile class fail, because we checked on
the argument being a str. Now we check on str or unicode, allowing both.

Note that in Python3, str is equivalent to unicode.

While having the possibility to read files with foreign characters
in their names is certainly a good thing, we recommend not to
use non-ascii characters in filenames, certainly not if you
want to use these names in external commands. And many things in
pyFormex will not yet work nicely with unicode strings.