Convert Process output to unicode
Since conversion to unicode seems to be the proper way to get the
output split in lines, both stdout and stderr are now converted.
A definition of unicode=str for Python3 (compat_3k) was added.

With these changes the pyFormex GUI can get launched under Python3
(though pyFormex is far from functional yet).