Applied the xrange fixer.
Just like zip and other functions, range becomes an iterator in Python3.
In many case range is used in a for... construct, so that poses no
problems. The cases where the full list is needed have been transformed
to either list(range()) or arange(). In most cases in pyFormex arange()
is the better option.