Cleanup sys.path
Now that all pyFormex imports have been sanitized to using absolute
paths, we have largely removed the sys.path manipulations that we did
in the past to allow import of modules.

Now the standard sys.path remains unchanged for packaged versions (source
or binary). When running directly from the source, the parent path
of the pyformex directory (containing the pyformex script and pyformex.conf
file) gets prepended to the path.