change avgNodalScalarOnAdjacentNodes into nodalAveraging
nodalAveraging (previously deprecated avgNodalScalarOnAdjacentNodes)
has been added and explained with an example.
It is probably doing something different than other mesh methods, as it
is operating on some data associated to the mesh (nodes), and
not on the mesh (geometry) itself.
Should this go to a new plugin (meshdata?) for data analysis
(e.g. for post-processing)??