Add a wait parameter to utils.Process (bug #40850)
The utils.Process class now has a wait parameter. The default wait=True
makes the caller wait for the subprocess to finish (which is what we
usually need). Setting wait=False lets the caller continue. You will
not be able to retrieve the stdout/stderr afterwards (well, not rightaway,
we may add some way to do it, but in general, when using wait=False, you
are not interested in the result of the command, just want to start some
process.

The timeout parameter of the Process.run method (and the whole run method)
has been incorporated in the Process initializer. Thus you should never
have to use the run method anymore.

In gui.draw.editFile, the parameter wait=False has been added. This should
fix bug #40850