vmtksurfacedistance: from sign from vmtk is fixed in pyFormex
the vmtkDistanceOfSurface and the vmtkDistanceOfPoints compute
the signed distance of some points from a surface.
The signed distance is positive if distance vector and normal
to the reference surface have negative dot product.
However, vmtk has a bug when the distance vector hits an edge.
This has been fixed in pyFormex by adding the findCell function,
which uses the vtkCellLocator