Restructure WebGL export
The export of WebGL has been completely overhauled. The WebGL class
is now in the first place intended as an exporter class, not as a
class to build a scene (It was needed at that time, but not anymore
in GL2 engine).

WebGL export is disabled with gl1 engine, we will now only support it
with gl2.

The following logic is used:
  W = WebGL()  # Create the exporter
  W.addScene()  # add the current gl2 scene
  W.exportScene() # export the scene
  W.export()  # finish the export

The two phases addScene and exportScene will later be merged into a single
call, and multiple models will be allowed:

  W = WebGL()
  W.addScene('scene1')
  # render a new scene
  W.addScene('scene2')
  # etc...
  W.export()